Cgi

Documentation

Types

type Cgi a = Replay Question Answer a

type Type = String

We use Type to differeniate between the types a user can enter into a textfield. We currently support int and str when type checking

type Id = Int

type Question = [(Id, String, Type)]

Question and Answer are the same type. The Id is unique for each question and is used later on to get the result. The answer itself is a String that might be able to be read as an Int if the Type is int. Not if the Type is str.

type Answer = Question

Operations

pickStr :: Int -> Answer -> String

Pick the string (answer value) from an answer

pickInt :: Int -> Answer -> Int

Pick the answer value and turn it to an int If the answer is "", just return 0

ask :: Read r => q -> Replay q r r

The ask fuction uses roughly the same process as liftIO.

If a question has already been answered in the trace, then we just return a Right, causing the program to go forward.

If the question hasn't been answered, we return a Left, causing the program to interrupt with a question

liftIO :: (Show a, Read a, Read r, Show r) => IO a -> Replay q r a

Lifting function to perform IO actions and store the result of them in our trace.

the first thing we do is checking the trace to see if an action has already been performed

If it has, then we just return it (running it through a helper function, r2a, to convert it to the appropriate type

If it hasn't, we pick (lift?) the IO action out from the IO monad, perform it and send the result to the trace

Run function

runCGI :: Cgi () -> IO ()

Run function for the CGI monad The trace is stored in one single hidden tag in the web page. That means we have to parse it when receiving it back from the web server.