Replay

Description

This module contains a replay monad. A replay monad can decide to stop in the middle of a running computation, and then later pick up where it left of.

We have a shallow implementation based around the semantics of the Replay type.

Documentation

Types

newtype Replay q r a

Type representing Replay

Trace r -> Trace r -> IO

Here, the first trace is used to do retracing. We pick Items from it until it's empty. When it's empty, retracing mode is complete.

The second trace is supposed to contain all traces from the start, and is then expanded with later computations

Constructors

Replay

Fields

runReplay :: Trace r -> Trace r -> IO (Either (q, Trace r) (a, Trace r))

data Item r

Constructors

Answer r

Result r

Operations

liftIO :: (Show a, Read a, Read r, Show r) => IO a -> Replay q r a

Lifting function to perform IO actions and store the result of them in our trace.

the first thing we do is checking the trace to see if an action has already been performed

If it has, then we just return it (running it through a helper function, r2a, to convert it to the appropriate type

If it hasn't, we pick (lift?) the IO action out from the IO monad, perform it and send the result to the trace

ask :: Read r => q -> Replay q r r

The ask fuction uses roughly the same process as liftIO.

If a question has already been answered in the trace, then we just return a Right, causing the program to go forward.

If the question hasn't been answered, we return a Left, causing the program to interrupt with a question

emptyTrace :: Trace r

addAnswer :: Trace r -> r -> Trace r

type Trace r = [Item r]

Type representing a trace of computations previously occured

Item defined what can happen in a computation

Run function

run :: (Show q, Show r, Show a) => Replay q r a -> Trace r -> IO (Either (q, Trace r) a)

Because we are using a shallow implementation, not muh work is done here. We feed the given replay with a trace and return the result of that run